+++ engine/lib/membrane_rtc_engine/track.ex
@@ -201,6 +201,6 @@ defmodule Membrane.RTC.Engine.Track do
   @doc """
   Checks whether track is a simulcast one or not.
   """
-  @spec is_simulcast?(t()) :: boolean()
-  def is_simulcast?(%__MODULE__{} = track), do: track.variants != [:high]
+  @spec simulcast?(t()) :: boolean()
+  def simulcast?(%__MODULE__{} = track), do: track.variants != [:high]
 end